# src/backend/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SVP Analytics API",
    "version": "1.0.0",
    "description": "SVP Partner Engagement Dashboard - Backend API (SRS v1.0)"
  },
  "servers": [
    {
      "url": "http://localhost:4001"
    }
  ],
  "paths": {
...
    "/api/chapters/": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "summary": "List all chapters",
        "description": "Returns a list of all chapters.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of chapters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/chapters/{id}": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "summary": "Get a chapter by ID",
        "description": "Returns a single chapter by its UUID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "format": "uuid",
            "description": "Chapter UUID"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chapter found"
          },
          "404": {
            "description": "Chapter not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}